+++ pytket/extensions/quantinuum/backends/quantinuum.py
@@ -17,7 +17,7 @@ from ast import literal_eval
 from dataclasses import dataclass
 import json
 from http import HTTPStatus
-from typing import Dict, List, Set, Optional, Sequence, Union, Any, cast
+from typing import Dict, List, Set, Optional, Sequence, Union, Any, cast, Tuple
 import warnings
 
 import numpy as np
@@ -778,6 +778,31 @@ class QuantinuumBackend(Backend):
                     handle, _convert_result(response["results"], ppcirc)
                 )
         return circ_status
+
+    def get_partial_result(
+        self, handle: ResultHandle
+    ) -> Tuple[Optional[BackendResult], CircuitStatus]:
+        """
+        Retrieve partial results for a given job, regardless of its current state.
+
+        :param handle: handle to results
+        :type handle: ResultHandle
+
+        :return: A tuple containing the results and circuit status. If no results are available, the first element is None.
+        :rtype: Tuple[Optional[BackendResult], CircuitStatus]
+        """
+        job_id = str(handle[0])
+        jr = self.api_handler.retrieve_job_status(job_id)
+        if not jr:
+            raise QuantinuumAPIError(f"Unable to retrive job {job_id}")
+        res = jr.get("results")
+        circ_status = _parse_status(jr)
+        if res is None:
+            return None, circ_status
+        ppcirc_rep = json.loads(cast(str, handle[1]))
+        ppcirc = Circuit.from_dict(ppcirc_rep) if ppcirc_rep is not None else None
+        backres = _convert_result(res, ppcirc)
+        return backres, circ_status
 
     def get_result(self, handle: ResultHandle, **kwargs: KwargTypes) -> BackendResult:
         """