+++ docs/changelog.rst
@@ -7,6 +7,7 @@ Changelog
 * cost function now takes the same kwargs as process_circuits
 * add check for the number of classical registers to the backend
 * Updated pytket version requirement to 1.14.1rc0
+* add ``get_partial_result`` method to ``QuantinuumBackend``.
 
 0.15.0 (April 2023)
 -------------------